dev: move action plan
