+++ pkg/model/action_plan.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package chi
+package model
 
 import (
 	"gopkg.in/d4l3k/messagediff.v1"
